Move test helpers into a separate module
